requirements.txt
aiogram==3.2.0
asyncpg
redis
python-dotenv
alembic
httpx
aiohttp
sqlalchemy
pydantic
fastapi
uvicorn
dataclass-rest
mulenpay_api